Added a search bar and item rarity images
-Added a search bar
-Added Item rarity images: Gold, Silver, and Copper (Bronze)
-Changed margin for filter-container
-Added header-container
-Changed --lightblue color opacity

Plan on every item having a rarity icon that the user can sort by. Kinda weird showing "hide gold items" if the user doesn't know which items are gold rarity.
Also plan on showing the item names in the boxes, but that may be difficult since some names are LONG.